requirements.txt
aiosqlite==0.17.0
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
ecdsa==0.17.0
fastapi==0.68.1
fastapi-pagination==0.8.3
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
iniconfig==1.1.1
iso8601==0.1.16
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.5
pytest-asyncio==0.15.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
uvicorn==0.15.0
